centralized atom in separate file
